Move fixFloat() into common.js.
